requirements.txt
fastapi
uvicorn
requests
python-dotenv
google-generativeai
faiss-cpu
sentence-transformers
streamlit